requirements.txt
# SCOPE-HN Dataset Processing Requirements

# Core dependencies
numpy>=1.21.0
pillow>=8.3.0
opencv-python>=4.5.0
matplotlib>=3.4.0

# Video processing
ffmpeg-python>=0.2.0

# Web interface for video trimming
flask>=2.0.0

# File handling and utilities
pathlib2>=2.3.0
tqdm>=4.62.0

# Data analysis and validation
pandas>=1.3.0

# Optional: For advanced image processing
scikit-image>=0.18.0

# Optional: For medical image handling
pydicom>=2.2.0

# Development and testing (optional)
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# Note: rclone must be installed separately for Google Drive access
# Installation: https://rclone.org/install/
